+++ src/dynamics/rigid_body.rs
@@ -157,7 +157,28 @@ impl Default for AdditionalMassProperties {
 /// and the `AdditionalMassProperties` should be modified instead).
 #[derive(Copy, Clone, Debug, Default, PartialEq, Component, Reflect)]
 #[reflect(Component, PartialEq)]
-pub struct ReadMassProperties(pub MassProperties);
+pub struct ReadMassProperties(MassProperties);
+
+impl ReadMassProperties {
+    /// Get the [`MassProperties`] of this rigid-body.
+    pub fn get(&self) -> &MassProperties {
+        &self.0
+    }
+
+    pub(crate) fn set(&mut self, mass_props: MassProperties) {
+        self.0 = mass_props;
+    }
+}
+
+/// Entity that likely had their mass properties changed this frame.
+#[derive(Deref, Copy, Clone, Debug, PartialEq, Event)]
+pub struct MassModifiedEvent(pub Entity);
+
+impl From<Entity> for MassModifiedEvent {
+    fn from(entity: Entity) -> Self {
+        Self(entity)
+    }
+}
 
 /// Center-of-mass, mass, and angular inertia.
 ///